+++ WalletWasabi.Fluent/ViewModels/SearchBar/SearchBarViewModel.cs
@@ -5,7 +5,7 @@ using System.Reactive.Linq;
 using DynamicData;
 using ReactiveUI;
 using WalletWasabi.Fluent.ViewModels.SearchBar.Patterns;
-using WalletWasabi.Fluent.ViewModels.SearchBar.SearchItem;
+using WalletWasabi.Fluent.ViewModels.SearchBar.SearchItems;
 
 namespace WalletWasabi.Fluent.ViewModels.SearchBar;
