+++ WalletWasabi.Fluent/ViewModels/SearchBar/SearchItems/AutocloseActionableItem.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using ReactiveUI;
 using WalletWasabi.Fluent.ViewModels.SearchBar.Patterns;
 
-namespace WalletWasabi.Fluent.ViewModels.SearchBar.SearchItem;
+namespace WalletWasabi.Fluent.ViewModels.SearchBar.SearchItems;
 
 public class AutocloseActionableItem : IActionableItem
 {
