+++ WalletWasabi.Fluent/ViewModels/SearchBar/SearchItems/IActionableItem.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 
-namespace WalletWasabi.Fluent.ViewModels.SearchBar.SearchItem;
+namespace WalletWasabi.Fluent.ViewModels.SearchBar.SearchItems;
 
 public interface IActionableItem : ISearchItem
 {
