+++ WalletWasabi.Fluent/ViewModels/SearchBar/SearchItems/ISearchItem.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using WalletWasabi.Fluent.ViewModels.SearchBar.Patterns;
 
-namespace WalletWasabi.Fluent.ViewModels.SearchBar.SearchItem;
+namespace WalletWasabi.Fluent.ViewModels.SearchBar.SearchItems;
 
 public interface ISearchItem
 {
