+++ WalletWasabi.Fluent/ViewModels/SearchBar/SearchItems/NonActionableSearchItem.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using WalletWasabi.Fluent.ViewModels.SearchBar.Patterns;
 
-namespace WalletWasabi.Fluent.ViewModels.SearchBar.SearchItem;
+namespace WalletWasabi.Fluent.ViewModels.SearchBar.SearchItems;
 
 public class NonActionableSearchItem : ISearchItem
 {
