+++ WalletWasabi.Fluent/ViewModels/SearchBar/Sources/ActionsSource.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using DynamicData;
 using WalletWasabi.Fluent.ViewModels.Navigation;
 using WalletWasabi.Fluent.ViewModels.SearchBar.Patterns;
-using WalletWasabi.Fluent.ViewModels.SearchBar.SearchItem;
+using WalletWasabi.Fluent.ViewModels.SearchBar.SearchItems;
 
 namespace WalletWasabi.Fluent.ViewModels.SearchBar.Sources;
