+++ WalletWasabi.Fluent/ViewModels/SearchBar/Sources/CompositeSearchItemsSource.cs
@@ -2,7 +2,7 @@ using System.Linq;
 using System.Reactive.Linq;
 using DynamicData;
 using WalletWasabi.Fluent.ViewModels.SearchBar.Patterns;
-using WalletWasabi.Fluent.ViewModels.SearchBar.SearchItem;
+using WalletWasabi.Fluent.ViewModels.SearchBar.SearchItems;
 
 namespace WalletWasabi.Fluent.ViewModels.SearchBar.Sources;
