+++ WalletWasabi.Fluent/ViewModels/SearchBar/Sources/ISearchItemSource.cs
@@ -1,6 +1,6 @@
 using DynamicData;
 using WalletWasabi.Fluent.ViewModels.SearchBar.Patterns;
-using WalletWasabi.Fluent.ViewModels.SearchBar.SearchItem;
+using WalletWasabi.Fluent.ViewModels.SearchBar.SearchItems;
 
 namespace WalletWasabi.Fluent.ViewModels.SearchBar.Sources;
